requirements.txt
# requirements.txt
# ===============================================================================
# CORE - Основные зависимости (НЕ МЕНЯЙ версии БЕЗ СОГЛАСОВАНИЯ!)
# если в процессе создания проекта ставишь новую библиотеку, добавь в этот файл
# и сообщи всем остальным разработчикам, сохрани на GitHub
# остальным для установки изменений выполнить pip install -r requirements.txt
# ================================================================================
# ==================================================
# CORE & ORCHESTRATION
# ==================================================
prefect==2.19.9             # Оркестратор пайплайнов, задачи, шедулер
chromadb==0.5.3             # Векторная база данных для LLM-памяти
rich==13.7.1                # Улучшенный вывод в консоль — цветной текст, таблицы, прогресс-бары
pydantic==2.8.2             # Валидация структур данных
loguru==0.7.2               # Удобное логирование
python-dotenv==1.0.1        # Загрузка переменных окружения (.env)

# ==================================================
# DATA & TRADING
# ==================================================
MetaTrader5==5.0.45         # Подключение к MT5, загрузка котировок
pandas==2.2.2               # Табличная обработка данных
numpy==1.26.4               # Математика, массивы, статистика

# ==================================================
# VISUALIZATION
# ==================================================
dash==2.17.1                # Веб-визуализация и интерактивные графики
plotly==5.24.0              # Основная библиотека для графиков
flask==3.0.3                # Веб-сервер для Dash
kaleido==0.2.1              # Экспорт графиков в PNG, PDF
gunicorn==21.2.0            # Продакшн сервер для Dash/Flask
streamlit==1.39.0           # UI интерфейс, мониторинг и контроль
streamlit-elements==0.1.0   # Компонент ReactFlow для Streamlit (визуал DAG схем)

# ==================================================
# AI / LLM INTEGRATION
# ==================================================
langchain==0.3.0            # Основной LLM фреймворк
openai==1.44.0              # Подключение GPT-4 / GPT-5
tiktoken==0.7.0             # Токенизация для OpenAI моделей
# anthropic==0.18.1         # (опционально) Claude API

# ==================================================
# INTEGRATIONS
# ==================================================
python-telegram-bot==21.9   # Telegram уведомления
notion-client==2.2.1        # Работа с Notion API
requests==2.32.3            # HTTP-запросы и REST API
pyyaml==6.0.2               # Работа с YAML конфигами

# ==================================================
# UTILITIES
# ==================================================
rich==13.7.1                # Улучшенный консольный вывод
python-dotenv==1.0.1        # Переменные окружения (.env)
loguru==0.7.2               # Логирование

# ==================================================
# DEVELOPMENT (опционально)
# ==================================================
# pytest==7.4.3             # Тестирование
# black==23.12.1            # Форматирование кода
# ruff==0.1.9               # Линтер
